network, events, and posts blueprints
